% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{add_hide_logo_css}
\alias{add_hide_logo_css}
\title{Add CSS rule to hide .logo class}
\usage{
add_hide_logo_css(file)
}
\arguments{
\item{file}{CSS file path}
}
\description{
Add CSS rule to hide .logo class
}
\keyword{internal}
