% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand_template.R
\name{brand_template}
\alias{brand_template}
\title{Example Brand Template}
\usage{
brand_template()
}
\value{
A [Brand] object.
}
\description{
Returns an example [Brand] object..
Includes color palette, Google and system fonts, and typography specs
for base text, headings, and links.

The template can be inspected as source code via [print_brand_template()],
and used as a starting point for custom branding passed to [create_website()].
}
\seealso{
[print_brand_template()], [Brand()], [create_website()]
}
