% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_s7.R
\name{BrandMeta}
\alias{BrandMeta}
\title{Brand Metadata}
\usage{
BrandMeta(name = character(0))
}
\arguments{
\item{name}{A character string that become the name inside the _brand.yml file}
}
\description{
Brand Metadata
}
