% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{callout_box}
\alias{callout_box}
\title{Create a callout box}
\usage{
callout_box(text, color, fixed_width_px = 300)
}
\arguments{
\item{text}{The text of the callout box. Character vector of length 1. Text that is supposed to be highlighted needs to be wrapped in <highlight></highlight> tags.}

\item{color}{The color of the callout box. One of the "version 600" colors from omni_colors(), i.e. "orange-red-600", "golden-yellow-600", "teal-600", "plum-600", "periwinkle-600"}

\item{fixed_width_px}{Width of the callout box in px. Must be numeric vector of length 1. Defaults to 300.}
}
\description{
Renders a callout box, dispatching to [callout_box_html()] or
[callout_box_typst()] depending on the output format (via
`knitr::is_html_output()`).
}
\seealso{
[callout_box_html()], [callout_box_typst()]
}
