% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{change_background_image}
\alias{change_background_image}
\title{Remove logo files in CSS}
\usage{
change_background_image(file, background_cover_image)
}
\arguments{
\item{file}{CSS file path}

\item{background_cover_image}{Path to the new background image}
}
\description{
Remove logo files in CSS
}
