% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{change_colors}
\alias{change_colors}
\title{Change main colors in CSS}
\usage{
change_colors(file, background_color = NULL, primary_font_color = NULL)
}
\arguments{
\item{file}{CSS file path}

\item{background_color}{Background color (optionnal)}

\item{primary_font_color}{Primary color (optionnal)}
}
\description{
Change main colors in CSS
}
