% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{change_fonts}
\alias{change_fonts}
\title{Change main fonts in CSS}
\usage{
change_fonts(file, main_font = NULL, secondary_font = NULL)
}
\arguments{
\item{file}{CSS file path}

\item{main_font}{New main font (optionnal)}

\item{secondary_font}{New secondary font (optionnal)}
}
\description{
Change main fonts in CSS
}
\keyword{internal}
