% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{change_to_csi_style}
\alias{change_to_csi_style}
\title{Use the CSI style}
\usage{
change_to_csi_style(file)
}
\arguments{
\item{file}{CSS file path}
}
\description{
Use the CSI style
}
\keyword{internal}
