% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{change_to_csi_style_html}
\alias{change_to_csi_style_html}
\title{Use the CSI style in HTML report}
\usage{
change_to_csi_style_html(file)
}
\arguments{
\item{file}{CSS file path}
}
\description{
Use the CSI style in HTML report
}
\keyword{internal}
