% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{change_to_csi_style_pdf}
\alias{change_to_csi_style_pdf}
\title{Use the CSI style in PDF report}
\usage{
change_to_csi_style_pdf(file)
}
\arguments{
\item{file}{CSS file path}
}
\description{
Use the CSI style in PDF report
}
\keyword{internal}
