% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_website.R
\name{create_website}
\alias{create_website}
\title{Create a new Quarto website}
\usage{
create_website(output_dir, brand = NULL, use_csi_logos = FALSE)
}
\arguments{
\item{output_dir}{New directory that will contain the Quarto
files.}

\item{brand}{Optional Brand object for custom branding
(created with `Brand()`. Default template available via `brand_template()`).
If not specified uses default _brand.yml file.}

\item{use_csi_logos}{Boolean that determines whether CSI logos will be used. Defaults to `FALSE`.}
}
\description{
`create_website()` creates a new directory with pre-defined
Quarto documents based on Omni branding.
}
