% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{footer_title_include}
\alias{footer_title_include}
\title{Build the hidden element that carries the running footer title}
\usage{
footer_title_include(footer_title)
}
\arguments{
\item{footer_title}{Title to use in the running footer}
}
\description{
Build the hidden element that carries the running footer title
}
\keyword{internal}
