% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{hide_acknowledgement_section}
\alias{hide_acknowledgement_section}
\title{Remove acknownledgement}
\usage{
hide_acknowledgement_section(file)
}
\arguments{
\item{file}{CSS file path}
}
\description{
Remove acknownledgement
}
\keyword{internal}
