% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{html_report}
\alias{html_report}
\title{Omni HTML Report}
\usage{
html_report(
  main_font = NULL,
  background_color = NULL,
  remove_logo = FALSE,
  use_csi_style = FALSE,
  ...
)
}
\arguments{
\item{main_font}{Main font for the report}

\item{background_color}{Background color}

\item{remove_logo}{Whether to remove Omni logo or not}

\item{use_csi_style}{Whether to use CSI style or not}

\item{...}{Params to pagedown::html_paged}
}
\value{
An rmd format
}
\description{
Omni HTML Report
}
