% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{icon_text}
\alias{icon_text}
\title{Text for Icon-Text-Grid}
\usage{
icon_text(text, width_px = 400)
}
\arguments{
\item{text}{Text that is placed ont the grid. Must be character vector of length 1. Can use Markdown notation.}

\item{width_px}{Width of the text. Must be numeric vector of length 1. Defaults to 400}
}
\description{
Renders text for the icon-text-grid, dispatching to [icon_text_html()] or
[icon_text_typst()] depending on the output format (via
`knitr::is_html_output()`).
}
\seealso{
[icon_text_html()], [icon_text_typst()]
}
