% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{icon_text_grid_typst}
\alias{icon_text_grid_typst}
\title{Place icons and texts on grid (Typst)}
\usage{
icon_text_grid_typst(..., width_px = 50, column_gap_px = 10, row_gap_px = 20)
}
\arguments{
\item{...}{Contents to be place on grid. Should consist of `omni_icon_html()` and `icon_text_html()` calls.}

\item{width_px}{Width of the circle in px. Numeric vector of length 1.}

\item{column_gap_px}{Gap between icon and text in px. Numeric vector of length 1.}

\item{row_gap_px}{Gap between rows in px. Numeric vector of length 1.}
}
\value{
Raw Typst markup, inserted via `knitr::asis_output()`.
}
\description{
This function creates the Typst markup that places the icons and texts from `omni_icon_typst()` and `icon_text_typst()` on a grid.
}
