% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{icon_text_typst}
\alias{icon_text_typst}
\title{Text for Icon-Text-Grid (Typst)}
\usage{
icon_text_typst(text, width_px = 400)
}
\arguments{
\item{text}{Text that is placed ont the grid. Must be character vector of length 1. Can use Markdown notation.}

\item{width_px}{Width of the text. Must be numeric vector of length 1. Defaults to 400}
}
\value{
Raw Typst markup, inserted via `knitr::asis_output()`.
}
\description{
This function creates the Typst markup that contains the text for the icon-text-grid
}
