% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_lato}
\alias{import_lato}
\title{Function to import the Lato font}
\usage{
import_lato()
}
\value{

}
\description{
This function imports the Lato font and makes it available
for use in R. You just need to run this function once on each computer.
Under the surface, this function uses fontextra::font_import(). For more
info on this process, please see
https://www.andrewheiss.com/blog/2017/09/27/working-with-r-cairo-graphics-custom-fonts-and-ggplot/
}
