% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{include_hypothesis}
\alias{include_hypothesis}
\title{Include Hypothesis in Rmd}
\usage{
include_hypothesis(add_hypothesis)
}
\arguments{
\item{add_hypothesis}{TRUE/FALSE}
}
\value{
Hypothesis included
}
\description{
Include Hypothesis in Rmd
}
