% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{knit_print.omni_table}
\alias{knit_print.omni_table}
\title{Print an Omni table in an R Markdown document}
\usage{
\method{knit_print}{omni_table}(x, ...)
}
\arguments{
\item{x}{An object created by `omni_table()`.}

\item{...}{Additional arguments passed to flextable's print method.}
}
\value{
The rendered table output.
}
\description{
Print an Omni table in an R Markdown document
}
