% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_report.R
\name{new_omni_report}
\alias{new_omni_report}
\title{Create a new report using the OMNI RMarkdown template}
\usage{
new_omni_report(file_name = "report")
}
\arguments{
\item{file_name}{The name of the file for your RMarkdown document. Defaults to "report" and creates a file called report.Rmd}
}
\value{

}
\description{
Create a new report using the OMNI RMarkdown template
}
