% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{number_emphasis_typst}
\alias{number_emphasis_typst}
\title{Create a number emphasis element (Typst)}
\usage{
number_emphasis_typst(
  number,
  text,
  color,
  font_size_pt = 16,
  text_font_size_pt = 12,
  fixed_width_px = 300
)
}
\arguments{
\item{number}{The emphasized number. Numeric or character vector of length 1.}

\item{text}{The info text. Character vector of length 1.}

\item{color}{Desired background color. Must be one `omni::omni_colors()`}

\item{font_size_pt}{Font size of emphasized number in pt. Numeric vector of length 1. Defaults to 16.}

\item{text_font_size_pt}{Font size of the info text in pt. Numeric vector of length 1. Defaults to 12.}

\item{fixed_width_px}{Width of the number emphasis in px. Must be numeric vector of length 1. Defaults to 300.}
}
\value{
Raw Typst markup for the desired number emphasis, inserted via `knitr::asis_output()`.
}
\description{
This function creates the Typst markup for the desired number emphasis.
}
