% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_formats.R
\name{omni_html_memo}
\alias{omni_html_memo}
\title{Omni HTML internal template}
\usage{
omni_html_memo(hypothesis = FALSE, ...)
}
\arguments{
\item{hypothesis}{Add Hypothesis}

\item{...}{Other params to bookdown::html_document2}
}
\value{
An rmd format
}
\description{
Omni HTML internal template
}
