% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_formats.R
\name{omni_html_report}
\alias{omni_html_report}
\title{Omni HTML Report to share with clients}
\usage{
omni_html_report(...)
}
\arguments{
\item{...}{Other params to bookdown::html_document2}
}
\value{
An rmd format
}
\description{
Omni HTML Report to share with clients
}
