% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_formats.R
\name{omni_html_slidy}
\alias{omni_html_slidy}
\title{Omni HTML Slidy template}
\usage{
omni_html_slidy(hypothesis = FALSE, ...)
}
\arguments{
\item{hypothesis}{Add Hypothesis}

\item{...}{Other params to rmarkdown::slidy_presentation}
}
\value{
An rmd format
}
\description{
Omni HTML Slidy template
}
