% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{omni_icon}
\alias{omni_icon}
\title{Get one of the pre-defined Omni icons}
\usage{
omni_icon(icon_name, width_px, icon_color_bg, icon_color_fg)
}
\arguments{
\item{icon_name}{The icon name. Must be character vector of length 1.}

\item{width_px}{Width of the circle in px. Numeric vector of length 1.}

\item{icon_color_bg}{Desired background color. Must be one `omni::omni_colors()`.}

\item{icon_color_fg}{Desired icon color. Should be 'white' or 'black' but technically all hex codes and CSS colors work.}
}
\description{
Renders an icon, dispatching to [omni_icon_html()] or [omni_icon_typst()]
depending on the output format (via `knitr::is_html_output()`).
}
\seealso{
[omni_icon_html()], [omni_icon_typst()]
}
