% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{omni_icon_html}
\alias{omni_icon_html}
\title{Get one of the pre-defined Omni icons (HTML)}
\usage{
omni_icon_html(icon_name, width_px, icon_color_bg, icon_color_fg)
}
\arguments{
\item{icon_name}{The icon name. Must be character vector of length 1.}

\item{width_px}{Width of the circle in px. Numeric vector of length 1.}

\item{icon_color_bg}{Desired background color. Must be one `omni::omni_colors()`.}

\item{icon_color_fg}{Desired icon color. Should be 'white' or 'black' but technically all hex codes and CSS colors work.}
}
\value{
HTML & CSS that of the desired icon (can be transformed into plain text using `as.character()`)
}
\description{
This function creates the HTML & CSS for the icon that should be placed into the text icon grid.
The HTML container of the icon always centers inside the surrounding environment.
}
\examples{
omni_icon_html('education', 50, 'teal-600', 'white')

}
