% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{omni_icons}
\alias{omni_icons}
\title{List all available Omni icons}
\usage{
omni_icons()
}
\value{
Character vector full of available omni icon names
}
\description{
This function creates the HTML & CSS for the desired number emphasis.
}
