% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_internal_simple.R
\name{omni_internal_simple}
\alias{omni_internal_simple}
\title{Omni internal RMD template}
\usage{
omni_internal_simple(hypothesis = FALSE, ...)
}
\arguments{
\item{...}{Other params to bookdown::html_document2}

\item{add_hypothesis}{Add Hypothesis}
}
\value{
An rmd format
}
\description{
Omni internal RMD template
}
