% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_formats.R
\name{omni_pdf_memo}
\alias{omni_pdf_memo}
\title{Omni Paged PDF Memo}
\usage{
omni_pdf_memo(main_font = NULL, header_font = NULL, main_color = NULL, ...)
}
\arguments{
\item{main_font}{Main font}

\item{header_font}{Header font}

\item{main_color}{Main color}

\item{...}{Other params to pagedown::html_paged}
}
\value{
An rmd format
}
\description{
Omni Paged PDF Memo
}
