% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_formats.R
\name{omni_pdf_report}
\alias{omni_pdf_report}
\title{Omni Paged PDF template}
\usage{
omni_pdf_report(...)
}
\arguments{
\item{...}{Other params to pagedown::html_paged}
}
\value{
An rmd format
}
\description{
Omni Paged PDF template
}
