% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_formats.R
\name{omni_pdf_report_new}
\alias{omni_pdf_report_new}
\title{Omni Paged PDF Report - new style}
\usage{
omni_pdf_report_new(...)
}
\arguments{
\item{...}{Other params to pagedown::html_paged}
}
\value{
An rmd format
}
\description{
Omni Paged PDF Report - new style
}
