% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_report.R
\name{omni_report}
\alias{omni_report}
\title{Create a new report using the OMNI RMarkdown template}
\usage{
omni_report(file_name = "report")
}
\arguments{
\item{file_name}{The name of the file for your RMarkdown document}
}
\value{

}
\description{
Create a new report using the OMNI RMarkdown template
}
