% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_formats.R
\name{omni_rmarkdown}
\alias{omni_rmarkdown}
\title{Omni HTML Rmarkdown template}
\usage{
omni_rmarkdown(...)
}
\arguments{
\item{...}{Other params to rmarkdown::html_document}
}
\value{
An rmd format
}
\description{
Omni HTML Rmarkdown template
}
