% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_rmd_format.R
\name{omni_rmd_template}
\alias{omni_rmd_template}
\title{Omni RMD template}
\usage{
omni_rmd_template(...)
}
\arguments{
\item{...}{Other params to pagedown::html_paged}
}
\value{
An rmd format
}
\description{
Omni RMD template
}
