% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{omni_set_page_start_pdf}
\alias{omni_set_page_start_pdf}
\title{Set the page counter in Omni pdf reports}
\usage{
omni_set_page_start_pdf(start_from, counter_start, total_pages = 100)
}
\arguments{
\item{start_from}{The page on which the counter should restart}

\item{counter_start}{The value of the first page}

\item{total_pages}{Deprecated. Kept for backward compatibility.}
}
\value{
An HTML character for rmarkdown
}
\description{
Controls the page counter for Omni PDF reports.
}
