% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_slidy.R
\name{omni_slidy}
\alias{omni_slidy}
\title{Omni Slidy template}
\usage{
omni_slidy(...)
}
\arguments{
\item{...}{Other params to rmarkdown::slidy_presentation}
}
\value{
An rmd format
}
\description{
Omni Slidy template
}
