% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_word.R
\name{omni_word}
\alias{omni_word}
\title{OMNI word output}
\usage{
omni_word(toc = FALSE, number_sections = TRUE, ...)
}
\arguments{
\item{toc}{Table of content as True/False}

\item{number_sections}{Number sections as true/false}

\item{...}{Other params to bookdown::word_document2}
}
\value{
An rmd format
}
\description{
OMNI word output
}
