% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{omnijs_run}
\alias{omnijs_run}
\title{Perform Omni Product Call}
\usage{
omnijs_run(args, type = c("omnigraffle", "omnioutliner"), wait = 0.1)
}
\arguments{
\item{args}{Script values to pass in.}

\item{type}{Program to use for automation. Default \code{"omnigraffle"}}

\item{wait}{Delay applied to function call before opening the URL.}
}
\value{
An opened session of OnmiGraffle or OmniOutliner with contents
posted from \emph{R}.
}
\description{
Perform Omni Product Call
}
\details{
There are two ways to structure URLs for omnigraffle.
\enumerate{
\item \code{"omnigraffle://localhost/omnijs-run?script=\%s"}
\item \code{"omnigraffle:///omnijs-run?script="}
}

Note, by adding a third \verb{/}, we are able to avoid typing \code{localhost}.
}
