% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_resource}
\alias{pkg_resource}
\title{Load resources - shortcut to system.file}
\usage{
pkg_resource(...)
}
\arguments{
\item{...}{Diverses arguments}
}
\description{
Load resources - shortcut to system.file
}
