% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_apply_labels}
\alias{qualtrics_apply_labels}
\title{Apply labels to survey}
\usage{
qualtrics_apply_labels(survey_data, labels_data)
}
\arguments{
\item{survey_data}{Output or subset of omni::qualtrics_get_survey()}

\item{labels_data}{Output of omni::qualtrics_get_labels()}
}
\value{
survey data with columns names replaced by labels
}
\description{
Apply labels to survey
}
