% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_get_labels}
\alias{qualtrics_get_labels}
\title{Get labels}
\usage{
qualtrics_get_labels(survey_id_name)
}
\arguments{
\item{survey_id_name}{Survey name or ID}
}
\value{
A dataframe with questions codes and labels
}
\description{
Get labels
}
