% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_get_survey}
\alias{qualtrics_get_survey}
\title{Get survey by name of ID}
\usage{
qualtrics_get_survey(survey_id_name, convert = FALSE, var_labels = FALSE, ...)
}
\arguments{
\item{survey_id_name}{Survey name or ID}

\item{convert}{Convert characters, default to FALSE}

\item{var_labels}{qualtRics::fetch_survey add_var_labels parameter, default to FALSE}

\item{...}{Others args}
}
\value{
A survey
}
\description{
Get survey by name of ID
}
