% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtromni.R
\name{qualtromni_connect}
\alias{qualtromni_connect}
\title{Connect to Qualtrics - need to have API key register before}
\usage{
qualtromni_connect(people_name)
}
\value{
Connection to Qualtrics
}
\description{
Connect to Qualtrics - need to have API key register before
}
