% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtromni.R
\name{qualtromni_get_survey}
\alias{qualtromni_get_survey}
\title{Get survey by name of ID}
\usage{
qualtromni_get_survey(survey_id_name)
}
\arguments{
\item{survey_id_name}{Survey name or ID, whatever}
}
\value{
A survey
}
\description{
Get survey by name of ID
}
