% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtromni.R
\name{qualtromni_register}
\alias{qualtromni_register}
\title{Write API key to the system - run only once}
\usage{
qualtromni_register(people_name, api_key, base_url)
}
\arguments{
\item{people_name}{Your name}

\item{api_key}{Your Qualtrics API key}
}
\value{
Env variables
}
\description{
Write API key to the system - run only once
}
