% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{quote_box}
\alias{quote_box}
\title{Create a quote box}
\usage{
quote_box(text, author, color, fixed_width_px = 300)
}
\arguments{
\item{text}{The text of the quote box. Character vector of length 1. Text that is supposed to be highlighted needs to be wrapped in <highlight></highlight> tags.}

\item{author}{The author of quote box. Character vector of length 1 or NULL (in case author line isn't required)}

\item{color}{The color of the quote box. One of the "version 600" colors from omni_colors(), i.e. "orange-red-600", "golden-yellow-600", "teal-600", "plum-600", "periwinkle-600"}

\item{fixed_width_px}{Width of the quote box in px. Must be numeric vector of length 1. Defaults to 300.}
}
\description{
Renders a quote box, dispatching to [quote_box_html()] or [quote_box_typst()]
depending on the output format (via `knitr::is_html_output()`).
}
\seealso{
[quote_box_html()], [quote_box_typst()]
}
