% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{read_renviron_qualtrics}
\alias{read_renviron_qualtrics}
\title{Read Renviron Qualtrics variables}
\usage{
read_renviron_qualtrics(organization_name = "OMNI")
}
\arguments{
\item{organization_name}{Organization name}
}
\value{
Environment variables loaded
}
\description{
Read Renviron Qualtrics variables
}
