% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{reduce_bottom_and_top_margin}
\alias{reduce_bottom_and_top_margin}
\title{Use the CSI style in HTML report}
\usage{
reduce_bottom_and_top_margin(file)
}
\arguments{
\item{file}{CSS file path}
}
\description{
Use the CSI style in HTML report
}
\keyword{internal}
