% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_cover_page}
\alias{remove_cover_page}
\title{Remove cover page from PDF}
\usage{
remove_cover_page(path)
}
\arguments{
\item{path}{Path to a PDF file}
}
\description{
Remove the first page of a PDF file.
}
\examples{
\dontrun{
remove_cover_page("my_file.pdf")

remove_cover_page("path/to/my/file.pdf")
}
}
