% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_cover_page_pdf}
\alias{remove_cover_page_pdf}
\title{Remove cover page from PDF}
\usage{
remove_cover_page_pdf(path)
}
\arguments{
\item{path}{Path to a PDF file}
}
\description{
Remove the first page of a PDF file.
}
\examples{
\dontrun{
remove_cover_page_pdf("my_file.pdf")

remove_cover_page_pdf("path/to/my/file.pdf")
}
}
