% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_omni_continuous}
\alias{scale_color_omni_continuous}
\title{Continuous color scale based on OMNI colors}
\usage{
scale_color_omni_continuous(palette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character vector of OMNI colors OR one of `c("orange-red", "golden-yellow", "olive-green", "plum", "teal", "periwinkle")`.}

\item{reverse}{Boolean, should palette be reversed?}

\item{...}{Additional arguments passed to scale_color_gradientn()}
}
\description{
Continuous color scale based on OMNI colors
}
\examples{
library(ggplot2)
ggplot(mpg, aes(x = displ, y = hwy, color = hwy)) +
    geom_point(size = 2.5) +
    scale_color_omni_continuous("orange-red")
}
