% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_periwinkle_omni_continuous}
\alias{scale_color_periwinkle_omni_continuous}
\title{Periwinkle continuous color scale based on OMNI colors}
\usage{
scale_color_periwinkle_omni_continuous(reverse = FALSE, ...)
}
\arguments{
\item{reverse}{Boolean, should palette be reversed?}

\item{...}{Additional arguments passed to `scale_color_gradientn()`}
}
\description{
Periwinkle continuous color scale based on OMNI colors
}
\examples{
library(ggplot2)
ggplot(mpg, aes(x = displ, y = hwy, color = hwy)) +
    geom_point(size = 2.5) +
    scale_color_periwinkle_omni_continuous()
}
