% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_omni.R
\name{scale_colour_omni_continuous}
\alias{scale_colour_omni_continuous}
\alias{scale_color_omni_continuous}
\title{Continuous color scale based on OMNI colors}
\usage{
scale_colour_omni_continuous(palette = "Main", reverse = FALSE, ...)

scale_color_omni_continuous(palette = "Main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in omni_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Continuous color scale based on OMNI colors
}
