% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_omni_continuous}
\alias{scale_fill_omni_continuous}
\title{Continuous fill scale based on OMNI colors}
\usage{
scale_fill_omni_continuous(palette = "Main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in omni_palettes ("Main" or "Blues")}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to scale_color_gradientn()}
}
\description{
Continuous fill scale based on OMNI colors
}
