% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_plum_omni_continuous}
\alias{scale_fill_plum_omni_continuous}
\title{Plum continuous fill scale based on OMNI colors}
\usage{
scale_fill_plum_omni_continuous(reverse = FALSE, ...)
}
\arguments{
\item{reverse}{Boolean, should palette be reversed?}

\item{...}{Additional arguments passed to scale_fill_gradientn()}

\item{palette}{Character vector of OMNI colors.}
}
\description{
Plum continuous fill scale based on OMNI colors
}
\examples{
library(ggplot2)
library(dplyr)

mpg_summary <- mpg |>
 group_by(class) |>
 summarise(avg_hwy = mean(hwy))

ggplot(mpg_summary, aes(x = class, y = avg_hwy, fill = avg_hwy)) +
  geom_bar(stat = "identity") +
  scale_fill_plum_omni_continuous()
}
