% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_rmd_format.R
\name{simple_report_format}
\alias{simple_report_format}
\title{Simple report format}
\usage{
simple_report_format(...)
}
\arguments{
\item{...}{Other params to pagedown::html_paged}
}
\value{
An rmd format
}
\description{
Simple report format
}
