% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{stage_css_assets}
\alias{stage_css_assets}
\title{Stage CSS assets in a writable temporary directory}
\usage{
stage_css_assets(file)
}
\arguments{
\item{file}{CSS file path}
}
\description{
Stage CSS assets in a writable temporary directory
}
\keyword{internal}
