% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{text_color}
\alias{text_color}
\title{Colorize text within inline text}
\usage{
text_color(text, color)
}
\arguments{
\item{text}{Must be a character vector of length 1.}

\item{color}{color to be used. Must be one of `omni_colors(named = TRUE)`}
}
\description{
Colorize text within inline text
}
