% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_client}
\alias{theme_client}
\title{Create theme for OMNI's clients.}
\usage{
theme_client(
  show_grid_lines = FALSE,
  show_legend = FALSE,
  base_family = "Inter Tight",
  plot_background_color = "White"
)
}
\arguments{
\item{show_grid_lines}{Whether or not to show grid lines. FALSE by default.}

\item{show_legend}{Whether or not to show the legend. FALSE by default.}

\item{base_family}{Base font family. Inter Tight by default.}

\item{plot_background_color}{Plot background color. White by default, can be set to Ivory.}
}
\value{
A ggplot2 theme
}
\description{
Create theme for OMNI's clients.
}
