% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_omni}
\alias{theme_omni}
\title{OMNI Institute ggplot2 theme}
\usage{
theme_omni(
  grid_lines = TRUE,
  base_theme = ggplot2::theme_minimal(base_family = "Lato")
)
}
\arguments{
\item{grid_lines}{Boolean to indicate whether to have grid_lines}

\item{base_theme}{The base theme used (theme_minimal() with the Lato font)}
}
\value{

}
\description{
OMNI Institute ggplot2 theme
}
\examples{
library(tidyverse)

iris \%>\%
  group_by(Species) \%>\%
  summarise(sepal_length_mean = mean(Sepal.Length)) \%>\%
  ggplot(aes(x = Species, y = sepal_length_mean, fill = Species)) +
  geom_bar(stat = "identity") +
  coord_flip() +
  theme_omni()
}
