% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{word_report}
\alias{word_report}
\title{Omni Word Report}
\usage{
word_report(...)
}
\arguments{
\item{...}{Other params to rmarkdown::word_document}
}
\value{
An rmd format
}
\description{
Omni Word Report
}
