% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_utils.R
\name{write_staged_css}
\alias{write_staged_css}
\title{Write modified CSS into a staged temporary directory}
\usage{
write_staged_css(file, css_lines)
}
\arguments{
\item{file}{CSS file path}

\item{css_lines}{CSS lines to write}
}
\description{
Write modified CSS into a staged temporary directory
}
\keyword{internal}
