% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a new Omni report}
\usage{
create_report(output_dir, format = c("pdf", "html"), use_csi_logos = FALSE)
}
\arguments{
\item{output_dir}{New directory that will contain the Quarto
report files.}

\item{format}{Report output format, either `"pdf"` or `"html"`.}

\item{use_csi_logos}{Boolean that determines whether CSI logos will be
used. Defaults to `FALSE`.}
}
\value{
The path to the created `template.qmd`, invisibly. In an
interactive session the file is also opened for editing.
}
\description{
`create_report()` creates a new directory with a pre-defined
Quarto report document (HTML or PDF) based on Omni branding.
}
