% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a new Omni report}
\usage{
create_report(
  output_dir,
  format = c("pdf", "html"),
  brand = NULL,
  use_csi_logos = FALSE
)
}
\arguments{
\item{output_dir}{New directory that will contain the Quarto
report files.}

\item{format}{Report output format, either `"html"` or `"pdf"`.}

\item{brand}{Optional Brand object for custom branding
(created with `Brand()`. Default template available via `brand_template()`).
If not specified uses default _brand.yml file.}

\item{use_csi_logos}{Boolean that determines whether CSI logos will be used. Defaults to `FALSE`.}
}
\description{
`create_report()` creates a new directory with a pre-defined
Quarto report document (HTML or PDF) based on Omni branding.
}
