% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_website.R
\name{create_website}
\alias{create_website}
\title{Create a new Quarto website}
\usage{
create_website(output_dir)
}
\arguments{
\item{output_dir}{New directory that will contain the Quarto
files.}
}
\description{
`create_website()` creates a new directory with pre-defined
Quarto documents based on Omni branding.
}
