% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{html_report}
\alias{html_report}
\title{Omni HTML Report}
\usage{
html_report(
  main_font = NULL,
  background_color = NULL,
  remove_logo = FALSE,
  ...
)
}
\arguments{
\item{...}{Params to pagedown::html_paged}
}
\value{
An rmd format
}
\description{
Omni HTML Report
}
