% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{icon_text}
\alias{icon_text}
\title{Text for Icon-Text-Grid}
\usage{
icon_text(text, width_px = 400)
}
\arguments{
\item{text}{Text that is placed ont the grid. Must be character vector of length 1. Can use Markdown notation.}

\item{width_px}{Width of the text. Must be numeric vector of length 1. Defaults to 400}
}
\value{
HTML & CSS
}
\description{
This function creates the HTML & CSS that contains the text for the icon-text-grid
}
