% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{number_emphasis}
\alias{number_emphasis}
\title{Create a number emphasis element}
\usage{
number_emphasis(number, text, color, fixed_width_px = 300)
}
\arguments{
\item{number}{The emphasized number. Numeric or character vector of length 1.}

\item{text}{The info text. Character vector of length 1.}

\item{color}{Desired background color. Must be one `omni::omni_colors()`}

\item{fixed_width_px}{Width of the number emphasis in px. Must be numeric vector of length 1. Defaults to 300.}
}
\value{
HTML & CSS that of the desired number emphasis
}
\description{
This function creates the HTML & CSS for the desired number emphasis.
}
