% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{omni_colors}
\alias{omni_colors}
\title{Function to extract OMNI colors as hex codes}
\usage{
omni_colors(...)
}
\arguments{
\item{...}{color or colors you want to return}
}
\value{

}
\description{
omni_colors() enables you to pull colors directly from the OMNI palette.
Choose one of the following options: Gray, Dark Gray, Dark Blue, Medium Blue,
Light Blue, Teal, Orange, or Tan. You'll use this function as follows:
omni_colors("Dark Blue") returns the hex value of the OMNI Dark Blue.
}
\examples{
omni_colors("Dark Blue", "Light Blue")

}
