% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{omni_colors}
\alias{omni_colors}
\title{Function to extract OMNI colors as hex codes}
\usage{
omni_colors(
  colors_sub = c("Dark Blue", "Medium Blue", "Light Blue", "Teal", "Orange", "Tan"),
  variations_sub = "0",
  named = FALSE
)
}
\arguments{
\item{colors_sub}{color or colors you want to return - vector}

\item{variations_sub}{Variation, default to 0 - vector}

\item{named}{Snould the output be named ? FALSE by default}
}
\value{
A color vector
}
\description{
omni_colors() enables you to pull colors directly from the OMNI palette.
Choose one of the following options: Gray, Medium Gray, Dark Gray, Dark Blue, Medium Blue,
Light Blue, Teal, Orange, or Tan. You'll use this function as follows:
omni_colors("Dark Blue") returns the hex value of the OMNI Dark Blue.
}
\examples{
omni_colors(c("Dark Blue", "Light Blue"))

}
