% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{omni_colors}
\alias{omni_colors}
\title{Function to extract WVC colors as hex codes}
\usage{
omni_colors(
  colors_sub = c("white", "ivory", "ivory.dark", "orange.red.light", "orange.red",
    "orange.red.dark", "golden.yellow.light", "golden.yellow", "golden.yellow.dark",
    "olive.green.light", "olive.green", "olive.green.dark", "teal.light", "teal",
    "teal.dark", "plum.light", "plum", "plum.dark", "periwinkle.light", "periwinkle",
    "periwinkle.dark", "steel.blue.light", "steel.blue.dark"),
  named = FALSE
)
}
\arguments{
\item{colors_sub}{Character vector of color names.}

\item{named}{Logical, should output be named? Default is FALSE.}
}
\value{
A vector of hex codes.
}
\description{
omni_colors() enables you to pull colors directly from the WVC palette.
Choose from:
"white", "ivory", "ivory.dark", "orange.red.light", "orange.red",
"orange.red.dark", "golden.yellow.light", "golden.yellow", "golden.yellow.dark",
"olive.green.light", "olive.green", "olive.green.dark", "teal.light", "teal",
"teal.dark", "plum.light", "plum", "plum.dark", "periwinkle.light", "periwinkle",
"periwinkle.dark", "steel.blue.light", "steel.blue.dark"
}
\examples{
omni_colors("orange.red")
omni_colors(c("plum", "golden.yellow", "teal"))
omni_colors(c("steel.blue.light", "olive.green"), named = TRUE)
}
