% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_omni.R
\name{omni_colors}
\alias{omni_colors}
\title{Function to extract OMNI colors as hex codes}
\usage{
omni_colors(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Function to extract OMNI colors as hex codes
}
