% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{omni_pal}
\alias{omni_pal}
\title{Function to interpolate a OMNI color palette}
\usage{
omni_pal(
  palette = c("orange.red", "golden.yellow", "olive.green", "teal", "plum", "periwinkle"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character vector of OMNI colors.}

\item{reverse}{Boolean, should palette be reversed?}

\item{...}{Additional arguments for colorRampPalette()}
}
\value{
A function that generates interpolated colors.
}
\description{
Function to interpolate a OMNI color palette
}
\examples{
omni_palette <- omni_pal(c("periwinkle.dark", "orange.red"))
omni_palette(10)
}
