% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{omni_pal}
\alias{omni_pal}
\title{Return function to interpolate an OMNI color palette}
\usage{
omni_pal(palette = "Main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in omni_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an OMNI color palette
}
