% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{omni_pal}
\alias{omni_pal}
\title{Function to interpolate a OMNI color palette}
\usage{
omni_pal(
  palette = c("orange-red 400", "golden-yellow 400", "olive-green 400", "teal 400",
    "plum 400", "periwinkle 400"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character vector of OMNI colors.}

\item{reverse}{Boolean, should palette be reversed?}

\item{...}{Additional arguments for colorRampPalette()}
}
\value{
A function that generates interpolated colors.
}
\description{
Function to interpolate a OMNI color palette
}
\examples{
omni_palette <- omni_pal(c("periwinkle 600", "orange-red 400"))
omni_palette(10)
}
