% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{omni_table}
\alias{omni_table}
\title{OMNI Table Function}
\usage{
omni_table(d)
}
\arguments{
\item{df}{The data frame to be put into the table}
}
\value{
A table themed
}
\description{
This function creates a table in OMNI's style.
}
