% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{omni_table}
\alias{omni_table}
\title{OMNI Table Function}
\usage{
omni_table(
  df,
  group_by = NULL,
  first_col_gray = FALSE,
  caption = NULL,
  with_stripes = TRUE
)
}
\arguments{
\item{df}{The data frame to be put into the table}

\item{group_by}{Character vector containing of grouping variables}

\item{first_col_gray}{Should the first column be gray. Default to FALSE}

\item{caption}{The caption of the table}

\item{with_stripes}{TRUE or FALSE depending on whether a striped pattern should be used. Defaults to TRUE (which uses stripes.)}
}
\value{
A table themed
}
\description{
This function creates a table in OMNI's style.
}
\examples{
palmerpenguins::penguins |>
 dplyr::slice(1:3) |>
   omni_table()

 palmerpenguins::penguins |>
   dplyr::slice(1:3) |>
   omni_table(first_col_gray = TRUE)

 palmerpenguins::penguins |>
   dplyr::slice(1:3, .by = species) |>
   omni_table(group_by = 'species')

 palmerpenguins::penguins |>
   dplyr::slice(1:3) |>
   omni_table(caption = 'Table 1. [Insert Table Name]')

# Without striped pattern
palmerpenguins::penguins |>
  dplyr::slice(1:3) |>
  omni_table(with_stripes = FALSE)

# Overwrite number formatting by transforming to character format
palmerpenguins::penguins |>
  dplyr::slice(1:3) |>
  dplyr::mutate(year = as.character(year)) |>
  omni_table()

}
