% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{omni_table}
\alias{omni_table}
\title{OMNI Table Function}
\usage{
omni_table(df, option = "flextable", grouped = NULL, first_col_blue = TRUE)
}
\arguments{
\item{df}{The data frame to be put into the table}

\item{option}{flextable (by default) or gt}

\item{grouped}{Only with option="flextable", use grouped output}

\item{first_col_blue}{Should the first column be blue. Default to TRUE}
}
\value{
A table themed
}
\description{
This function creates a table in OMNI's style.
}
