% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_formats.R
\name{omni_word_report}
\alias{omni_word_report}
\title{OMNI Word Template}
\usage{
omni_word_report(number_sections = TRUE, ...)
}
\arguments{
\item{number_sections}{Number sections as true/false}

\item{...}{Other params to bookdown::word_document2}
}
\value{
An rmd format
}
\description{
OMNI Word Template
}
