% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{pdf_report}
\alias{pdf_report}
\title{Omni Paged PDF Report}
\usage{
pdf_report(
  main_font = NULL,
  secondary_font = NULL,
  background_color = NULL,
  primary_color = NULL,
  remove_logo = FALSE,
  remove_cover_page = FALSE,
  ...
)
}
\arguments{
\item{...}{Other params to pagedown::html_paged}
}
\value{
An rmd format
}
\description{
Omni Paged PDF Report
}
