% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{pdf_report}
\alias{pdf_report}
\title{Omni Paged PDF Report}
\usage{
pdf_report(
  main_font = NULL,
  secondary_font = NULL,
  background_cover_image = NULL,
  background_color = NULL,
  primary_font_color = NULL,
  remove_logo = FALSE,
  remove_cover_page = FALSE,
  remove_title_page = FALSE,
  remove_toc_page = FALSE,
  ...
)
}
\arguments{
\item{main_font}{Main font}

\item{secondary_font}{Secondary font}

\item{background_cover_image}{Image to use for the background in the cover page. It must be one of (case insensitive): `c("01-yellow", "02-teal", "03-orangered", "06-teal", "07-periwinkle", "07-olive", "08-plum")`.}

\item{background_color}{Background color of the document}

\item{primary_font_color}{Primary color, mostly used in titles.}

\item{remove_logo}{Whether to remove Omni logos from the document.}

\item{remove_cover_page}{Whether to remove the cover page.}

\item{remove_title_page}{Whether to remove the title page.}

\item{remove_toc_page}{Whether to remove the TOC (table of content) page.}

\item{...}{Additional arguments passed to `pagedown::html_paged()`}
}
\value{
An rmd format
}
\description{
Omni Paged PDF Report
}
