% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_connect}
\alias{qualtrics_connect}
\title{Connect to Qualtrics - need to have API key register before}
\usage{
qualtrics_connect(people_name)
}
\value{
Connection to Qualtrics
}
\description{
Connect to Qualtrics - need to have API key register before
}
