% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_connect}
\alias{qualtrics_connect}
\title{Connect to Qualtrics}
\usage{
qualtrics_connect(organization_name = "OMNI", api_key = "", base_url = "")
}
\arguments{
\item{organization_name}{Organization name}

\item{api_key}{Your Qualtrics API key}

\item{base_url}{Base URL}
}
\value{
Connection to Qualtrics
}
\description{
This function connect to Qualtrics. It calls omni::qualtrics_register behind the scenes
to provide a connection if no API key and URL are registered (or if they are expired).
In that case, please provide `api_key` and `base_url`
}
