% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_get_survey}
\alias{qualtrics_get_survey}
\title{Get survey by name of ID}
\usage{
qualtrics_get_survey(survey_id_name)
}
\arguments{
\item{survey_id_name}{Survey name or ID}
}
\value{
A survey
}
\description{
Get survey by name of ID
}
