% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_register}
\alias{qualtrics_register}
\title{Write API key to the system - run only once}
\usage{
qualtrics_register(organization_name = "OMNI", api_key, base_url)
}
\arguments{
\item{organization_name}{Name of your organization (e.g. OMNI)}

\item{api_key}{Your Qualtrics API key}

\item{base_url}{Base URL}
}
\value{
Env variables
}
\description{
Write API key to the system - run only once
}
