% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elements.R
\name{quote_box_html}
\alias{quote_box_html}
\title{Create a quote box (HTML)}
\usage{
quote_box_html(text, author, color, fixed_width_px = 300)
}
\arguments{
\item{text}{The text of the quote box. Character vector of length 1. Text that is supposed to be highlighted needs to be wrapped in <highlight></highlight> tags.}

\item{author}{The author of quote box. Character vector of length 1 or NULL (in case author line isn't required)}

\item{color}{The color of the quote box. One of the "version 600" colors from omni_colors(), i.e. "orange-red-600", "golden-yellow-600", "teal-600", "plum-600", "periwinkle-600"}

\item{fixed_width_px}{Width of the quote box in px. Must be numeric vector of length 1. Defaults to 300.}
}
\value{
HTML & CSS that of the desired quote box
}
\description{
This function creates the HTML & CSS for the desired quote boxes. It can use the "version 600" colors of the color palette and highlight specific text via `<highlight>` tags.
}
\examples{

quote_text <- paste0(
  'This is a quote box. You can <highlight>change text color to ',
  'highlight certain parts</highlight>, or just leave the text all ',
  'white. Change the background color as desired to match the page.'
)

htmltools::browsable(
  quote_box_html(
    text = quote_text,
    author = 'John Jacob, random guy',
    color = 'olive-green-600'
  )
)

htmltools::browsable(
  quote_box_html(
    text = quote_text,
    author = 'John Jacob, random guy',
    color = 'periwinkle-600'
  )
)

}
