% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_omni_discrete}
\alias{scale_color_omni_discrete}
\title{Discrete color scale based on OMNI colors}
\usage{
scale_color_omni_discrete(palette = "Main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in omni_palettes (Main or Blues)}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Discrete color scale based on OMNI colors
}
