% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_omni_discrete}
\alias{scale_color_omni_discrete}
\title{Discrete color scale based on OMNI colors}
\usage{
scale_color_omni_discrete(
  palette = c("orange.red", "golden.yellow", "olive.green", "teal", "plum", "periwinkle"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character vector of OMNI colors.}

\item{reverse}{Boolean, should palette be reversed?}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Discrete color scale based on OMNI colors
}
\examples{
library(ggplot2)
ggplot(mpg, aes(x = displ, y = hwy, color = class)) +
    geom_point(size = 2.5) +
    scale_color_omni_discrete()
}
