% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_omni_continuous}
\alias{scale_fill_omni_continuous}
\title{Continuous fill scale based on OMNI colors}
\usage{
scale_fill_omni_continuous(
  palette = c("orange-red 200", "orange-red 400", "orange-red 600"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character vector of OMNI colors.}

\item{reverse}{Boolean, should palette be reversed?}

\item{...}{Additional arguments passed to scale_fill_gradientn()}
}
\description{
Continuous fill scale based on OMNI colors
}
\examples{
library(ggplot2)
library(dplyr)

mpg_summary <- mpg |>
 group_by(class) |>
 summarise(avg_hwy = mean(hwy))

ggplot(mpg_summary, aes(x = class, y = avg_hwy, fill = avg_hwy)) +
  geom_bar(stat = "identity") +
  scale_fill_omni_continuous()
}
