% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_omni_discrete}
\alias{scale_fill_omni_discrete}
\title{Discrete fill scale based on OMNI colors}
\usage{
scale_fill_omni_discrete(
  palette = c("orange-red 400", "golden-yellow 400", "olive-green 400", "teal 400",
    "plum 400", "periwinkle 400"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character vector of OMNI colors.}

\item{reverse}{Boolean, should palette be reversed?}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Discrete fill scale based on OMNI colors
}
\examples{
library(ggplot2)
ggplot(mpg, aes(x = class, fill = class)) +
  geom_bar() +
  scale_fill_omni_discrete()
}
