% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_defaults.R
\name{set_omni_defaults}
\alias{set_omni_defaults}
\title{Update defaults to OMNI's theme}
\usage{
set_omni_defaults(base_family = "Inter Tight", base_color = "#405065")
}
\arguments{
\item{base_family}{The base font family for the theme.}

\item{base_color}{Base color}
}
\description{
Update defaults to OMNI's theme
}
