% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_omni}
\alias{theme_omni}
\title{OMNI Institute ggplot2 theme}
\usage{
theme_omni(show_grid_lines = TRUE, show_legend = TRUE)
}
\arguments{
\item{show_legend}{Whether or not to show the legend (TRUE by default)}

\item{grid_lines}{Boolean to indicate whether to have grid_lines (TRUE by default)}
}
\value{

}
\description{
OMNI Institute ggplot2 theme
}
