% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_omni}
\alias{theme_omni}
\title{OMNI Institute ggplot2 theme}
\usage{
theme_omni(
  show_grid_lines = TRUE,
  show_legend = TRUE,
  base_theme = ggplot2::theme_minimal(base_family = "Lato")
)
}
\arguments{
\item{show_legend}{Whether or not to show the legend (TRUE by default)}

\item{base_theme}{The base theme used (theme_minimal() with the Lato font)}

\item{grid_lines}{Boolean to indicate whether to have grid_lines (TRUE by default)}
}
\value{

}
\description{
OMNI Institute ggplot2 theme
}
