% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_omni}
\alias{theme_omni}
\title{OMNI Institute ggplot2 theme}
\usage{
theme_omni(show_grid_lines = TRUE, show_legend = TRUE, base_family = "Calibri")
}
\arguments{
\item{show_grid_lines}{Show grid lines or not}

\item{show_legend}{Whether or not to show the legend (TRUE by default)}

\item{base_family}{Base font family (Lato by default). If you want to use a custom font, you need to register it first.}
}
\description{
OMNI Institute ggplot2 theme
}
