% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_omni}
\alias{theme_omni}
\title{OMNI Institute ggplot2 theme}
\usage{
theme_omni(
  grid_lines = TRUE,
  base_theme = ggplot2::theme_minimal(base_family = "Lato")
)
}
\arguments{
\item{base_theme}{}
}
\value{

}
\description{
OMNI Institute ggplot2 theme
}
